Rename backend nodes (#171)
* renamed backend_input node to network_input

* renamed backend_output node to robot_communication

* fixed formatting
