+++ src/thunderbots/software/robot_communication/grsim/grsim_backend.h
@@ -4,10 +4,10 @@
 #include <string>
 
 #include "ai/world/team.h"
-#include "backend_output/backend.h"
 #include "geom/angle.h"
 #include "geom/point.h"
 #include "proto/grSim_Packet.pb.h"
+#include "robot_communication/backend.h"
 
 
 class GrSimBackend : public Backend
