+++ src/thunderbots/software/robot_communication/main.cpp
@@ -7,8 +7,8 @@
 
 #include "ai/primitive/move_primitive.h"
 #include "ai/primitive/primitive.h"
-#include "backend_output/grsim/grsim_backend.h"
 #include "geom/point.h"
+#include "robot_communication/grsim/grsim_backend.h"
 #include "util/constants.h"
 #include "util/logger/init.h"
 
@@ -37,7 +37,7 @@ void primitiveUpdateCallback(const thunderbots_msgs::PrimitiveArray::ConstPtr& m
 int main(int argc, char** argv)
 {
     // Init ROS node
-    ros::init(argc, argv, "backend_output");
+    ros::init(argc, argv, "robot_communication");
     ros::NodeHandle node_handle;
 
     // Create subscribers to topics we care about
