+++ src/thunderbots/software/test/grsim_backend/grsim_backend.cpp
@@ -1,13 +1,13 @@
-#include "backend_output/grsim/grsim_backend.h"
+#include "robot_communication/grsim/grsim_backend.h"
 
 #include <google/protobuf/util/message_differencer.h>
 #include <gtest/gtest.h>
 
 #include <limits>
 
-#include "backend_output/backend.h"
 #include "proto/grSim_Commands.pb.h"
 #include "proto/grSim_Packet.pb.h"
+#include "robot_communication/backend.h"
 
 TEST(GrSimBackendTest, create_grsim_packet_zero_vel)
 {
